[FlexNG] Bottom margin when fragmenting
The margin bottom of a flex item is applied as a result
of the stored block-offset of each item, unless it was the last
item in the column or an item in the last row. If the last item
had a bottom margin, this margin wouldn't be added to the intrinsic
block size.

To fix this, store the margin bottom in the flex item and add it
to the intrinsic block size when the item has completed layout
for its own content.

Note that multi-line-column-flex-fragmentation-050.html
was already passing before this CL since a block-size
must be set on the flex container for the items to wrap,
which constrains the margins of the last items.

Bug: 1339487
Change-Id: I7099d6bab7e6a18d003ccea302d6573a5e45782b