WhitespaceChildrenMayChange needs to be handled for all elements
We tried to optimize away WhitespaceChildrenMayChange handling for
subtrees which should regardlessly be re-attached, but descendants
further down could still be marked for re-attachment which caused us to
stop early for ancestor marking at those whitespace-marked elements.

Remove the ReattachLayoutTree() check and handle the whitespace flag if
there is a LayoutObject.

Bug: 1338779
Change-Id: Iad95fdd674ca679137e66d5d4394efda0a8a4571